requirements.txt
pytz==2021.1
pysolar==0.9
python_dateutil==2.8.1
PyYAML==5.4.1
schedule==1.0.0
